+++ library/src/main/java/com/loopj/android/http/ResponseHandlerInterface.java
@@ -53,7 +53,7 @@ public interface ResponseHandlerInterface {
      * @param bytesWritten number of written bytes
      * @param bytesTotal   number of total bytes to be written
      */
-    void sendProgressMessage(int bytesWritten, int bytesTotal);
+    void sendProgressMessage(long bytesWritten, long bytesTotal);
 
     /**
      * Notifies callback, that request was cancelled
@@ -148,7 +148,7 @@ public interface ResponseHandlerInterface {
      * This method is called once by the system when the response is about to be
      * processed by the system. The library makes sure that a single response
      * is pre-processed only once.
-     *
+     * <p/>
      * Please note: pre-processing does NOT run on the main thread, and thus
      * any UI activities that you must perform should be properly dispatched to
      * the app's UI thread.
@@ -162,7 +162,7 @@ public interface ResponseHandlerInterface {
      * This method is called once by the system when the request has been fully
      * sent, handled and finished. The library makes sure that a single response
      * is post-processed only once.
-     *
+     * <p/>
      * Please note: post-processing does NOT run on the main thread, and thus
      * any UI activities that you must perform should be properly dispatched to
      * the app's UI thread.
